+++ packages/expo-router/src/getRoutesSSR.ts
@@ -17,7 +17,7 @@ export type Options = Omit<OptionsCore, 'getSystemRoute'>;
  */
 export function getRoutes(contextModule: RequireContext, options: Options = {}): RouteNode | null {
   return getRoutesCore(contextModule, {
-    getSystemRoute({ route, type, defaults }) {
+    getSystemRoute({ route, type, defaults, redirectConfig, rewriteConfig }) {
       if (route === '' && type === 'layout') {
         // Root layout when no layout is defined.
         return {
@@ -58,15 +58,26 @@ export function getRoutes(contextModule: RequireContext, options: Options = {}):
           dynamic: [{ name: '+not-found', deep: true, notFound: true }],
           children: [],
         };
-      } else if ((type === 'redirect' || type === 'rewrite') && defaults) {
+      } else if (type === 'redirect' && redirectConfig && defaults) {
         return {
           ...defaults,
           loadRoute() {
-            return require('./getRoutesRedirects').getRedirectModule(route);
+            return require('./getRoutesRedirects').getRedirectModule(redirectConfig);
+          },
+        };
+      } else if (type === 'rewrite' && rewriteConfig && defaults) {
+        return {
+          ...defaults,
+          loadRoute() {
+            return {
+              default: contextModule(rewriteConfig.destinationContextKey).default,
+            };
           },
         };
       }
-      throw new Error(`Unknown system route: ${route} and type: ${type}`);
+      throw new Error(
+        `Unknown system route: ${route} and type: ${type} and redirectConfig: ${redirectConfig} and rewriteConfig: ${rewriteConfig}`
+      );
     },
     ...options,
   });